% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_normal.R,
%   R/functions_multiple_tte.R
\name{dbivanorm}
\alias{dbivanorm}
\title{Density of the bivariate normal distribution}
\usage{
dbivanorm(x, y, mu1, mu2, sigma1, sigma2, rho)

dbivanorm(x, y, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{x}{integral variable}

\item{y}{integral variable}

\item{mu1}{mean of second endpoint}

\item{mu2}{mean of first endpoint}

\item{sigma1}{standard deviation of first endpoint}

\item{sigma2}{standard deviation of second endpoint}

\item{rho}{correlation between the two endpoints}
}
\value{
The Function \code{dbivanorm()} will return the density of a bivariate normal distribution.
}
\description{
Density of the bivariate normal distribution
}
\keyword{internal}
