% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugdevelopResult.R
\name{drugdevelopResult}
\alias{drugdevelopResult}
\title{Construct a drugdevelopResult object from a data frame}
\usage{
drugdevelopResult(x, ...)
}
\arguments{
\item{x}{data frame}
}
\value{
the same data frame, equipped with class "drugdevelopResult"
}
\description{
This is a short wrapper for adding the "drugdevelopR" string to the list
of S3 classes of a data frame.
}
\keyword{internal}
