% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multitrial_normal.R
\name{Epgo23_normal}
\alias{Epgo23_normal}
\title{Expected probability to do third phase III trial: Epgo3}
\usage{
Epgo23_normal(kappa, n2, w, Delta1, Delta2, in1, in2, a, b)
}
\arguments{
\item{kappa}{threshold value for the go/no-go decision rule}

\item{n2}{total sample size for phase II; must be even number}

\item{w}{weight for mixture prior distribution}

\item{Delta1}{assumed true treatment effect for standardized difference in means}

\item{Delta2}{assumed true treatment effect for standardized difference in means}

\item{in1}{amount of information for Delta1 in terms of sample size}

\item{in2}{amount of information for Delta2 in terms of sample size}

\item{a}{lower boundary for the truncation}

\item{b}{upper boundary for the truncation}
}
\value{
the output of the the function Epgo23_normal is the expected number of participants in phase III with conservative decision rule and sample size calculation:  use lower bound of one-sided confidence intervall
}
\description{
Expected probability to do third phase III trial: Epgo3
}
\examples{
res <- Epgo23_normal(kappa = 0.1, n2 = 50, w = 0.3,
                               Delta1 = 0.375, Delta2 = 0.625, in1 = 300, in2 = 600, 
                               a = 0.25, b = 0.75)
}
