% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multitrial_binary.R
\name{EPsProg3_binary}
\alias{EPsProg3_binary}
\title{Expected probability of a successful program: EsP}
\usage{
EPsProg3_binary(
  RRgo,
  n2,
  alpha,
  beta,
  p0,
  w,
  p11,
  p12,
  in1,
  in2,
  case,
  size,
  fixed
)
}
\arguments{
\item{RRgo}{threshold value for the go/no-go decision rule}

\item{n2}{total sample size for phase II; must be even number}

\item{alpha}{significance level}

\item{beta}{1-beta power for calculation of sample size for phase III}

\item{p0}{assumed true rate of control group}

\item{w}{weight for mixture prior distribution}

\item{p11}{assumed true rate of treatment group}

\item{p12}{assumed true rate of treatment group}

\item{in1}{amount of information for p11 in terms of sample size}

\item{in2}{amount of information for p12 in terms of sample size}

\item{case}{choose case: "at least 1, 2 or 3 significant trials needed for approval"}

\item{size}{size category "small", "medium" or "large"}

\item{fixed}{choose if true treatment effects are fixed or random}
}
\value{
the output of the the function EPsProg3_binary is the expected probability of a successful program
}
\description{
Expected probability of a successful program: EsP
}
\examples{
res <- EPsProg3_binary(RRgo = 0.8, n2 = 50,  alpha = 0.025, beta = 0.1, 
                                 p0 = 0.6  w = 0.3, p11 =  0.3, p12 = 0.5, 
                                 in1 = 300, in2 = 600, case = 2, size = "small",
                                 fixed = FALSE)
}
