% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_bias_binary.R
\name{EPsProg_binary_R2}
\alias{EPsProg_binary_R2}
\title{Expected probability of a successful program: EsP}
\usage{
EPsProg_binary_R2(
  RRgo,
  n2,
  Adj,
  alpha,
  beta,
  step1,
  step2,
  p0,
  w,
  p11,
  p12,
  in1,
  in2,
  fixed
)
}
\arguments{
\item{RRgo}{threshold value for the go/no-go decision rule}

\item{n2}{total sample size for phase II; must be even number}

\item{Adj}{adjustment parameter}

\item{alpha}{significance level}

\item{beta}{1-beta power for calculation of sample size for phase III}

\item{step1}{lower boundary for effect size}

\item{step2}{upper boundary for effect size}

\item{p0}{assumed true rate of control group}

\item{w}{weight for mixture prior distribution}

\item{p11}{assumed true rate of treatment group}

\item{p12}{assumed true rate of treatment group}

\item{in1}{amount of information for p11 in terms of sample size}

\item{in2}{amount of information for p12 in terms of sample size}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE Delta1 is used as fixed effect}
}
\value{
the output of the the function EPsProg_binary_R2 is the expected probability of a successful program
}
\description{
Expected probability of a successful program: EsP
}
\examples{
res <- EPsProg_binary_R2(RRgo = 0.8, n2 = 50, Adj = 0, alpha = 0.025, beta = 0.1, 
                                 step1 = 1, step2 = 0.95, p0 = 0.6  w = 0.3,
                                 p11 =  0.3, p12 = 0.5, in1 = 300, in2 = 600, 
                                 fixed = FALSE)
}
