% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_normal.R
\name{EPsProg_multiple_normal}
\alias{EPsProg_multiple_normal}
\title{expected probability of a successful program}
\usage{
EPsProg_multiple_normal(
  kappa,
  n2,
  alpha,
  beta,
  Delta1,
  Delta2,
  sigma1,
  sigma2,
  step11,
  step12,
  step21,
  step22,
  in1,
  in2,
  fixed,
  rho
)
}
\arguments{
\item{kappa}{threshold value for the go/no-go decision rule; vector for both endpoints}

\item{n2}{total sample size for phase II; must be even number}

\item{alpha}{significance level}

\item{beta}{1-beta power for calculation of sample size for phase III}

\item{Delta1}{assumed true treatment effect for standardized difference in means}

\item{Delta2}{assumed true treatment effect for standardized difference in means}

\item{sigma1}{standard deviation of first endpoint}

\item{sigma2}{standard deviation of second endpoint}

\item{step11}{lower boundary for effect size for first endpoint}

\item{step12}{lower boundary for effect size for second endpoint}

\item{step21}{upper boundary for effect size for first endpoint}

\item{step22}{upper boundary for effect size for second endpoint}

\item{in1}{amount of information for Delta1 in terms of sample size}

\item{in2}{amount of information for Delta2 in terms of sample size}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE Delta1 is used as fixed effect}

\item{rho}{correlation between the two endpoints}
}
\value{
the output of the the function EPsProg_multiple_normal is the expected probability of a successfull program, when going to phase III
}
\description{
expected probability of a successful program
}
\examples{
res <- EPsProg_multiple_normal(kappa = c(0.1,0.1), n2 = 50, alpha = 0.025, beta = 0.1,
                               Delta1 = 0.375, Delta2 = 0.625, sigma1 = 8, sigma2 = 4,
                               step11 = 0, step12 = 0, step21 = 0.5, step22 = 0.5, 
                               in1 = 300, in2 = 600, fixed = FALSE, rho = 0.3)
}
