% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_bias_normal.R
\name{EPsProg_normal_L}
\alias{EPsProg_normal_L}
\title{Expected probability of a successful program: EsP}
\usage{
EPsProg_normal_L(
  kappa,
  n2,
  Adj,
  alpha,
  beta,
  step1,
  step2,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  fixed
)
}
\arguments{
\item{kappa}{threshold value for the go/no-go decision rule}

\item{n2}{total sample size for phase II; must be even number}

\item{Adj}{adjustment parameter}

\item{alpha}{significance level}

\item{beta}{1-beta power for calculation of sample size for phase III}

\item{step1}{lower boundary for effect size}

\item{step2}{upper boundary for effect size}

\item{w}{weight for mixture prior distribution}

\item{Delta1}{assumed true treatment effect for standardized difference in means}

\item{Delta2}{assumed true treatment effect for standardized difference in means}

\item{in1}{amount of information for Delta1 in terms of sample size}

\item{in2}{amount of information for Delta2 in terms of sample size}

\item{a}{lower boundary for the truncation}

\item{b}{upper boundary for the truncation}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE Delta1 is used as fixed effect}
}
\value{
the output of the the function EPsProg_normal_L is the expected probability of a successful program
}
\description{
Expected probability of a successful program: EsP
}
\examples{
res <- EPsProg_normal_L(kappa = 0.1, n2 = 50, Adj = 0, alpha = 0.025, beta = 0.1, w = 0.3,
                                 step1 = 0, step2 = 0.5,
                                 Delta1 = 0.375, Delta2 = 0.625, in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, fixed = FALSE)
}
