% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_tte.R
\name{EPsProg_tte}
\alias{EPsProg_tte}
\title{expected probability of a successful program}
\usage{
EPsProg_tte(
  HRgo,
  d2,
  alpha,
  beta,
  step1,
  step2,
  w,
  hr1,
  hr2,
  id1,
  id2,
  gamma,
  fixed
)
}
\arguments{
\item{HRgo}{threshold value for the go/no-go decision rule;}

\item{alpha}{significance level}

\item{beta}{1-beta power for calculation of sample size for phase III}

\item{step1}{lower boundary for effect size}

\item{step2}{upper boundary for effect size}

\item{hr1}{assumed true treatment effect on HR scale for treatment 1}

\item{hr2}{assumed true treatment effect on HR scale for treatment 2}

\item{id1}{amount of information for hr1 in terms of sample size}

\item{id2}{amount of information for hr2 in terms of sample size}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE hr1 is used as fixed effect}

\item{n2}{total sample size for phase II; must be even number}

\item{ec}{control arm event rate for phase II and III}

\item{rho}{correlation between the two endpoints}
}
\value{
the output of the the function EPsProg_tte is the expected probability of a successfull program, when going to phase III
}
\description{
expected probability of a successful program
}
\examples{
res <- EPsProg_tte(HRgo = 0.8, n2 = 50, alpha = 0.025, beta = 0.1,
                               ec = 0.6, hr1 = 0.75, hr2 = 0.80,
                               id1 = 300, id2 = 600, 
                               step1 = 1, step2 = 0.95,
                               fixed = FALSE, rho = 0.3)
}
