% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_normal.R
\name{get_sample_multiple_normal}
\alias{get_sample_multiple_normal}
\title{Generate sample for Monte Carlo integration in the multiple setting}
\usage{
get_sample_multiple_normal(Delta1, Delta2, in1, in2, rho)
}
\description{
Generate sample for Monte Carlo integration in the multiple setting
}
\keyword{internal}
