% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_tte.R
\name{get_sample_multiple_tte}
\alias{get_sample_multiple_tte}
\title{Generate sample for Monte Carlo integration in the multiple setting}
\usage{
get_sample_multiple_tte(hr1, hr2, id1, id2, rho)
}
\description{
Generate sample for Monte Carlo integration in the multiple setting
}
\keyword{internal}
