% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_bias_binary.R
\name{optimal_bias_binary}
\alias{optimal_bias_binary}
\title{Optimal phase II/III drug development planning when discounting phase II results with binary endpoint}
\usage{
optimal_bias_binary(
  w,
  p0,
  p11,
  p12,
  in1,
  in2,
  n2min,
  n2max,
  stepn2,
  rrgomin,
  rrgomax,
  steprrgo,
  adj = "both",
  lambdamin = NULL,
  lambdamax = NULL,
  steplambda = NULL,
  alphaCImin = NULL,
  alphaCImax = NULL,
  stepalphaCI = NULL,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K = Inf,
  N = Inf,
  S = -Inf,
  steps1 = 1,
  stepm1 = 0.95,
  stepl1 = 0.85,
  b1,
  b2,
  b3,
  fixed = FALSE,
  num_cl = 1
)
}
\arguments{
\item{w}{weight for mixture prior distribution}

\item{p0}{assumed true rate of control group}

\item{p11}{assumed true rate of treatment group}

\item{p12}{assumed true rate of treatment group}

\item{in1}{amount of information for p11 in terms of sample size}

\item{in2}{amount of information for p12 in terms of sample size}

\item{n2min}{minimal total sample size for phase II; must be even number}

\item{n2max}{maximal total sample size for phase II, must be even number}

\item{stepn2}{stepsize for the optimization over n2; must be even number}

\item{rrgomin}{minimal threshold value for the go/no-go decision rule}

\item{rrgomax}{maximal threshold value for the go/no-go decision rule}

\item{steprrgo}{stepsize for the optimization over RRgo}

\item{lambdamin}{minimal adjustment parameter lambda}

\item{lambdamax}{maximal adjustment parameter lambda}

\item{steplambda}{stepsize for the adjustment parameter lambda}

\item{alphaCImin}{minimal alphaCI}

\item{alphaCImax}{maximal alphaCI}

\item{stepalphaCI}{stepsize for alphaCI}

\item{alpha}{significance level}

\item{beta}{1-beta power for calculation of sample size for phase III}

\item{c2}{variable per-patient cost for phase II}

\item{c3}{variable per-patient cost for phase III}

\item{c02}{fixed cost for phase II}

\item{c03}{fixed cost for phase III}

\item{K}{constraint on the costs of the program, default: Inf, e.g. no constraint}

\item{N}{constraint on the total expected sample size of the program, default: Inf, e.g. no constraint}

\item{S}{constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint}

\item{steps1}{lower boundary for effect size category "small" in RR scale, default: 1}

\item{stepm1}{lower boundary for effect size category "medium" in RR scale = upper boundary for effect size category "small" in RR scale, default: 0.95}

\item{stepl1}{lower boundary for effect size category "large" in RR scale = upper boundary for effect size category "medium" in RR scale, default: 0.85}

\item{b1}{expected gain for effect size category "small"}

\item{b2}{expected gain for effect size category "medium"}

\item{b3}{expected gain for effect size category "large"}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE p11 is used as fixed effect for p1}

\item{num_cl}{number of clusters used for parallel computing, default: 1}

\item{gamma}{to model different populations in phase II and III choose gamma!=0, default: 0}
}
\value{
The output of the function \code{\link{optimal_binary}} is a data.frame containing the optimization results:
\describe{
\item{Method}{Type of adjustment: multipl. (multiplicative) or add. (additive)}
\item{u}{maximal expected utility}
\item{Adj}{optimal adjustment parameter (lambda or alphaCI according to Method)}
\item{RRgo}{optimal threshold value for the decision rule to go to phase III}
\item{n2}{total sample size for phase II}
\item{n3}{total sample size for phase III; rounded to the next even natural number}
\item{n}{total sample size in the program; n = n2 + n3}
\item{K}{maximal costs of the program}
\item{pgo}{probability to go to phase III}
\item{sProg}{probability of a successful program}
\item{sProg1}{probability of a successful program with "small" treatment effect in Phase III}
\item{sProg2}{probability of a successful program with "medium" treatment effect in Phase III}
\item{sProg3}{probability of a successful program with "large" treatment effect in Phase III }
\item{K2}{expected costs for phase II}
\item{K3}{expected costs for phase III}
}
and further input parameters.

Taking cat(comment()) of the data.frame object lists the used optimization sequences, start and finish date of the optimization procedure.
}
\description{
The drugdevelopR package enables planning of phase II/III drug development programs with optimal sample size allocation and go/no-go decision rules including methods for discounting of phase II results (Preussler et. al, 2020).. For binary endpoints the treatment effect is measured by the risk ratio (RR).The assumed true treatment effects can be assumed fixed or modelled by a prior distribution. The R Shiny application \href{https://web.imbi.uni-heidelberg.de/prior/}{prior} visualizes the prior distributions used in this package. Fast coputing is enabled by parallel programming.
}
\section{drugdevelopR functions}{

The drugdevelopR package provides the functions
\itemize{
\item \code{\link{optimal_tte}},
\item \code{\link{optimal_binary}} or
\item \code{\link{optimal_normal}}
}
to plan optimal phase II/III drug development programs with
\itemize{
\item time-to-event (treatment effect measured by hazard ratio (HR)),
\item binary (treatment effect measured by risk ratio (RR)) and
\item normally distributed (treatment effect measured by standardized difference in means (Delta))
}
endpoint, where the treatment effect is modelled by a \href{https://web.imbi.uni-heidelberg.de/prior/}{prior}. Optimal phase II/III drug development planning with fixed treatment effects can be done with the help of the R Shiny application \href{https://web.imbi.uni-heidelberg.de/basic/}{basic}. Extensions are
\itemize{
\item optimal planning of programs including methods for discounting of phase II results (function: \code{\link{optimal_bias}}, App: \href{https://web.imbi.uni-heidelberg.de/bias/}{bias}),
\item optimal planning of programs with several phase III trials (function: \code{\link{optimal_multitrial}}, App: \href{https://web.imbi.uni-heidelberg.de/multitrial/}{multitrial}) and
\item optimal planning of programs with multiple arms (function: \code{\link{optimal_multiarm}}, App: \href{https://web.imbi.uni-heidelberg.de/multiarm/}{multiarm}).
}
}

\examples{
res <- optimal_bias_binary(w = 0.3,                           # define parameters for prior
  p0 = 0.6, p11 =  0.3, p12 = 0.5, in1 = 30, in2 = 60,   # (https://web.imbi.uni-heidelberg.de/prior/)
  n2min = 20, n2max = 100, stepn2 = 4,                   # define optimization set for n2
  rrgomin = 0.7, rrgomax = 0.9, steprrgo = 0.05,         # define optimization set for RRgo
  adj = "both",                                          # choose type of adjustment
  alpha = 0.05, beta = 0.1,                              # drug development planning parameters
  lambdamin = 0.2, lambdamax = 1, steplambda = 0.05,     # define optimization set for lambda
  alphaCImin = 0.025, alphaCImax = 0.5, stepalphaCI = 0.025, # define optimization set for alphaCI
  c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,               # define fixed and variable costs for phase II and III,
  K = Inf, N = Inf, S = -Inf,                            # set maximal costs/ expected sample size for the program or minimal expected probability of a successful program
  steps1 = 1,                                            # define lower boundary for "small"
  stepm1 = 0.95,                                         # "medium"
  stepl1 = 0.85,                                         # and "large" treatment effect size categories as proposed by IQWiG (2016)
  b1 = 1000, b2 = 2000, b3 = 3000,                       # define expected benefit for a "small", "medium" and "large" treatment effect
  fixed = FALSE,                                         # choose if true treatment effects are fixed or random
  num_cl = 1)                                            # set number of cores used for parallelized computing (check maximum number possible with detectCores())
res
cat(comment(res))                                        # displays the optimization sequence, start and finish date of the optimization procedure.
}
\references{
IQWiG (2016). Allgemeine Methoden. Version 5.0, 10.07.2016, Technical Report. Available at \href{https://www.iqwig.de/de/methoden/methodenpapier.3020.html}{https://www.iqwig.de/de/methoden/methodenpapier.3020.html}, assessed last 15.05.19.
}
\section{Last Edited By}{
Johannes Cepicka
}

\section{Last Edit Date}{
2022-04-23
}

