% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_binary.R
\name{optimal_binary}
\alias{optimal_binary}
\title{Optimal phase II/III drug development planning with binary endpoint}
\usage{
optimal_binary(
  w,
  p0,
  p11,
  p12,
  in1,
  in2,
  n2min,
  n2max,
  stepn2,
  rrgomin,
  rrgomax,
  steprrgo,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K = Inf,
  N = Inf,
  S = -Inf,
  steps1 = 1,
  stepm1 = 0.95,
  stepl1 = 0.85,
  b1,
  b2,
  b3,
  gamma = 0,
  fixed = FALSE,
  skipII = FALSE,
  num_cl = 1
)
}
\arguments{
\item{w}{weight for \href{https://web.imbi.uni-heidelberg.de/prior/}{mixture prior distribution}}

\item{p0}{assumed true rate of control group, see
\href{https://web.imbi.uni-heidelberg.de/prior/}{here}
for details}

\item{p11}{assumed true rate of treatment group, see
\href{https://web.imbi.uni-heidelberg.de/prior/}{here}
for details}

\item{p12}{assumed true rate of treatment group, see
\href{https://web.imbi.uni-heidelberg.de/prior/}{here}
for details}

\item{in1}{amount of information for \code{p11} in terms of sample size, see
\href{https://web.imbi.uni-heidelberg.de/prior/}{here}
for details}

\item{in2}{amount of information for \code{p12} in terms of sample size, see
\href{https://web.imbi.uni-heidelberg.de/prior/}{here}
for details}

\item{n2min}{minimal total sample size for phase II; must be an even number}

\item{n2max}{maximal total sample size for phase II, must be an even number}

\item{stepn2}{step size for the optimization over n2; must be an even number}

\item{rrgomin}{minimal threshold value for the go/no-go decision rule}

\item{rrgomax}{maximal threshold value for the go/no-go decision rule}

\item{steprrgo}{step size for the optimization over RRgo}

\item{alpha}{significance level}

\item{beta}{type II error rate; i.e. \code{1 - beta} is the power for calculation of the number of events for phase III}

\item{c2}{variable per-patient cost for phase II in 10^5 $}

\item{c3}{variable per-patient cost for phase III in 10^5 $}

\item{c02}{fixed cost for phase II in 10^5 $}

\item{c03}{fixed cost for phase III in 10^5 $}

\item{K}{constraint on the costs of the program, default: Inf, e.g. no constraint}

\item{N}{constraint on the total expected sample size of the program, default: Inf, e.g. no constraint}

\item{S}{constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint}

\item{steps1}{lower boundary for effect size category "small" in RR scale, default: 1}

\item{stepm1}{lower boundary for effect size category "medium" in RR scale = upper boundary for effect size category "small" in RR scale, default: 0.95}

\item{stepl1}{lower boundary for effect size category "large" in RR scale = upper boundary for effect size category "medium" in RR scale, default: 0.85}

\item{b1}{expected gain for effect size category "small"}

\item{b2}{expected gain for effect size category "medium"}

\item{b3}{expected gain for effect size category "large"}

\item{gamma}{to model different populations in phase II and III choose \code{gamma != 0}, default: 0, see
\href{https://web.imbi.uni-heidelberg.de/prior/}{here}
for details}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE p11 is used as fixed effect for p1}

\item{skipII}{skipII choose if skipping phase II is an option, default: FALSE;
if TRUE, the program calculates the expected utility for the case when phase
II is skipped and compares it to the situation when phase II is not skipped.
The results are then returned as a list of two results lists, \code{res[[1]]}
being the results when including phase II and \code{res[[2]]} when skipping phase II.}

\item{num_cl}{number of clusters used for parallel computing, default: 1}
}
\value{
The output of the function is a \code{data.frame} object containing the optimization results:
\describe{\item{u}{maximal expected utility under the optimization constraints, i.e. the expected utility of the optimal sample size and threshold value}\item{RRgo}{optimal threshold value for the decision rule to go to phase III}\item{n2}{total sample size for phase II; rounded to the next even natural number}
\item{n3}{total sample size for phase III; rounded to the next even natural number}
\item{n}{total sample size in the program; n = n2 + n3}
\item{K}{maximal costs of the program}
\item{pgo}{probability to go to phase III}
\item{sProg}{probability of a successful program}
\item{sProg1}{probability of a successful program with "small" treatment effect in phase III}
\item{sProg2}{probability of a successful program with "medium" treatment effect in phase III}
\item{sProg3}{probability of a successful program with "large" treatment effect in phase III}
\item{K2}{expected costs for phase II}
\item{K3}{expected costs for phase III}}
and further input parameters. Taking \code{cat(comment())} of the
data frame lists the used optimization sequences, start and
finish date of the optimization procedure.
}
\description{
The \code{\link{optimal_binary}} function of the drugdevelopR package enables
planning of phase II/III drug development programs with optimal sample size
allocation and go/no-go decision rules for binary endpoints. In this case,
the treatment effect is measured by the risk ratio (RR). The assumed true
treatment effects can be assumed to be fixed or modelled by a prior
distribution. The R Shiny application
\href{https://web.imbi.uni-heidelberg.de/prior/}{prior} visualizes the prior
distributions used in this package. Fast computing is enabled by parallel
programming.
}
\examples{
res <- optimal_binary(w = 0.3,                           # define parameters for prior
  p0 = 0.6, p11 =  0.3, p12 = 0.5, in1 = 30, in2 = 60,   # (https://web.imbi.uni-heidelberg.de/prior/)
  n2min = 20, n2max = 100, stepn2 = 4,                   # define optimization set for n2
  rrgomin = 0.7, rrgomax = 0.9, steprrgo = 0.05,         # define optimization set for RRgo
  alpha = 0.05, beta = 0.1,                              # drug development planning parameters
  c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,               # define fixed and variable costs for phase II and III,
  K = Inf, N = Inf, S = -Inf,                            # set maximal costs/ expected sample size for the program or minimal expected probability of a successful program
  steps1 = 1,                                            # define lower boundary for "small"
  stepm1 = 0.95,                                         # "medium"
  stepl1 = 0.85,                                         # and "large" treatment effect size categories as proposed by IQWiG (2016)
  b1 = 1000, b2 = 2000, b3 = 3000,                       # define expected benefit for a "small", "medium" and "large" treatment effect
  gamma = 0,                                             # assume different/same population structures in phase II and III
  fixed = FALSE,                                         # choose if true treatment effects are fixed or random
  skipII = FALSE,                                        # choose if skipping phase II would be an option
  num_cl = 1)                                            # set number of cores used for parallelized computing (check maximum number possible with detectCores())
res
cat(comment(res))                                        # displays the optimization sequence, start and finish date of the optimization procedure.

}
\references{
IQWiG (2016). Allgemeine Methoden. Version 5.0, 10.07.2016, Technical Report. Available at \href{https://www.iqwig.de/de/methoden/methodenpapier.3020.html}{https://www.iqwig.de/de/methoden/methodenpapier.3020.html}, assessed last 15.05.19.
}
