% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_multiple_generic.R
\name{optimal_multiple_generic}
\alias{optimal_multiple_generic}
\title{Generic function for optimizing drug development programs with multiple endpoints}
\usage{
optimal_multiple_generic(rho, alpha, n2min, n2max, stepn2, fixed)
}
\arguments{
\item{rho}{correlation between the two endpoints}

\item{alpha}{one-sided significance level/family-wise error rate}

\item{n2min}{minimal total sample size in phase II, must be divisible by 3}

\item{n2max}{maximal total sample size in phase II, must be divisible by 3}

\item{stepn2}{stepsize for the optimization over n2, must be divisible by 3}

\item{fixed}{assumed fixed treatment effect}
}
\value{
No return value
}
\description{
This function is only used for documentation generation.
}
\keyword{internal}
