% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_multiple_normal.R
\name{optimal_multiple_normal}
\alias{optimal_multiple_normal}
\title{Optimal phase II/III drug development planning for programs with multiple
normally distributed endpoints}
\usage{
optimal_multiple_normal(
  Delta1,
  Delta2,
  in1,
  in2,
  sigma1,
  sigma2,
  n2min,
  n2max,
  stepn2,
  kappamin,
  kappamax,
  stepkappa,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K = Inf,
  N = Inf,
  S = -Inf,
  steps1 = 0,
  stepm1 = 0.5,
  stepl1 = 0.8,
  b1,
  b2,
  b3,
  rho,
  fixed,
  relaxed = FALSE,
  num_cl = 1
)
}
\arguments{
\item{Delta1}{assumed true treatment effect for endpoint 1 measured as the
standardized difference in means}

\item{Delta2}{assumed true treatment effect for endpoint 2 measured as the
standardized difference in means}

\item{in1}{amount of information for Delta1 in terms of number of events}

\item{in2}{amount of information for Delta2 in terms of number of events}

\item{sigma1}{variance of endpoint 1}

\item{sigma2}{variance of endpoint 1}

\item{n2min}{minimal total sample size in phase II, must be divisible by 3}

\item{n2max}{maximal total sample size in phase II, must be divisible by 3}

\item{stepn2}{stepsize for the optimization over n2, must be divisible by 3}

\item{kappamin}{minimal threshold value kappa for the go/no-go decision rule}

\item{kappamax}{maximal threshold value  kappa for the go/no-go decision rule}

\item{stepkappa}{step size for the optimization over the threshold value kappa}

\item{alpha}{one-sided significance level/family-wise error rate}

\item{beta}{type-II error rate for any pair, i.e. \code{1 - beta} is the (any-pair) power for calculation of the sample size for phase III}

\item{c2}{variable per-patient cost for phase II in 10^5 $}

\item{c3}{variable per-patient cost for phase III in 10^5 $}

\item{c02}{fixed cost for phase II in 10^5 $}

\item{c03}{fixed cost for phase III in 10^5 $}

\item{K}{constraint on the costs of the program, default: Inf, e.g. no constraint}

\item{N}{constraint on the total expected sample size of the program, default: Inf, e.g. no constraint}

\item{S}{constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint}

\item{steps1}{lower boundary for effect size category "small", default: 0}

\item{stepm1}{lower boundary for effect size category "medium" = upper boundary for effect size category "small" default: 0.5}

\item{stepl1}{lower boundary for effect size category "large" = upper boundary for effect size category "medium", default: 0.8}

\item{b1}{expected gain for effect size category "small" in 10^5 $}

\item{b2}{expected gain for effect size category "medium" in 10^5 $}

\item{b3}{expected gain for effect size category "large" in 10^5 $}

\item{rho}{correlation between the two endpoints}

\item{fixed}{assumed fixed treatment effect}

\item{relaxed}{relaxed or strict decision rule}

\item{num_cl}{number of clusters used for parallel computing, default: 1}
}
\value{
The output of the function is a \code{data.frame} object containing the optimization results:
\describe{\item{u}{maximal expected utility under the optimization constraints, i.e. the expected utility of the optimal sample size and threshold value}\item{Kappa}{optimal threshold value for the decision rule to go to phase III}\item{n2}{total sample size for phase II; rounded to the next even natural number}
\item{n3}{total sample size for phase III; rounded to the next even natural number}
\item{n}{total sample size in the program; n = n2 + n3}
\item{K}{maximal costs of the program}
\item{pgo}{probability to go to phase III}
\item{sProg}{probability of a successful program}
\item{sProg1}{probability of a successful program with "small" treatment effect in phase III}
\item{sProg2}{probability of a successful program with "medium" treatment effect in phase III}
\item{sProg3}{probability of a successful program with "large" treatment effect in phase III}
\item{K2}{expected costs for phase II}
\item{K3}{expected costs for phase III}}
and further input parameters. Taking \code{cat(comment())} of the
data frame lists the used optimization sequences, start and
finish date of the optimization procedure.
}
\description{
The function \code{\link{optimal_multiple_normal}} of the drugdevelopR
package enables planning of phase II/III drug development programs with
optimal sample size allocation and go/no-go decision rules for two normally
distributed endpoints and one control group. (Preussler et. al, 2019).
Planning is also possible via user friendly R Shiny App:
\href{https://web.imbi.uni-heidelberg.de/multiarm/}{multiarm}.
Fast computing is enabled by parallel programming.
}
\examples{
 #res <- optimal_multiple_normal(Delta1 = 0.75, Delta2 = 0.80,    # define assumed true HRs
 # in1=300, in2=600, sigma1 = 8, sigma2= 12,
 # n2min = 30, n2max = 90, stepn2 = 6,                    # define optimization set for n2
 # kappamin = 0.7, kappamax = 0.9, stepkappa = 0.05,         # define optimization set for HRgo
 # alpha = 0.05, beta = 0.1,                              # drug development planning parameters
 # c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,               # define fixed and variable costs for phase II and III
 # K = Inf, N = Inf, S = -Inf,                            # set maximal costs/ expected sample size for the program or minimal expected probability of a successful program
 # steps1 = 0,                                            # define lower boundary for "small"
 # stepm1 = 0.5,                                          # "medium"
 # stepl1 = 0.8,                                         # and "large" treatment effect size categories as proposed by IQWiG (2016)
 #  b1 = 1000, b2 = 2000, b3 = 3000,                       # define expected benefit for a "small", "medium" and "large" treatment effect
 # rho = 0.5, relaxed = TRUE,                             # relaxed "TRUE"
 # fixed = TRUE,                                          #   treatment effect
 # num_cl = 1)                                            # set number of cores used for parallelized computing (check maximum number possible with detectCores())
#res
#cat(comment(res))                                        # displays the optimization sequence, start and finish date of the optimization procedure.

}
\references{
Preussler, S., Kirchner, M., Goette, H., Kieser, M. (2019). Optimal Designs for Multi-Arm Phase II/III Drug Development Programs. Submitted to peer-review journal.

IQWiG (2016). Allgemeine Methoden. Version 5.0, 10.07.2016, Technical Report. Available at \href{https://www.iqwig.de/de/methoden/methodenpapier.3020.html}{https://www.iqwig.de/de/methoden/methodenpapier.3020.html}, assessed last 15.05.19.
}
