% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_multitrial_generic.R
\name{optimal_multitrial_generic}
\alias{optimal_multitrial_generic}
\title{Generic function for optimizing multi-trial programs}
\usage{
optimal_multitrial_generic(case, strategy = TRUE)
}
\arguments{
\item{case}{choose case: "at least 1, 2 or 3 significant trials needed for approval"}

\item{strategy}{choose strategy: "conduct 1, 2, 3 or 4 trials in order to achieve the case's goal"; TRUE calculates all strategies of the selected \code{case}}
}
\description{
Generic function for optimizing multi-trial programs
}
\section{Effect sizes}{

In other settings, the definition of "small", "medium" and "large" effect
sizes can be user-specified using the input parameters \code{steps1}, \code{stepm1} and
\code{stepl1}. Due to the complexity of the multitrial setting, this feature is
not included for this setting. Instead, the effect sizes were set to
to predefined values as explained under sProg1, sProg2 and sProg3 in the
\emph{Value} section.
}

\keyword{internal}
