% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_return_doc.R
\name{optimal_return_doc}
\alias{optimal_return_doc}
\title{Function for generating documentation of return values}
\usage{
optimal_return_doc(type, setting = "basic")
}
\arguments{
\item{type}{string deciding whether this is return text for
normal, binary or time-to-event endpoints}

\item{setting}{string containing the setting, i.e. "basic", "bias",
"multitrial"}
}
\value{
string containing the documentation of the return value.
}
\description{
Generates the documentation of the return value
of an optimal function including some custom text.
}
\keyword{internal}
