% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_tte.R
\name{optimal_tte}
\alias{optimal_tte}
\title{Optimal phase II/III drug development planning with time-to-event endpoint}
\format{data.frame containing the optimization results (see Value)}
\usage{
optimal_tte(w, hr1, hr2, id1, id2, d2min, d2max, stepd2, hrgomin, hrgomax,
  stephrgo, alpha, beta, xi2, xi3, c2, c3, c02, c03, K = Inf, N = Inf,
  S = -Inf, steps1 = 1, stepm1 = 0.95, stepl1 = 0.85, b1, b2, b3,
  gamma = 0, fixed = FALSE, skipII = FALSE, num_cl = 1)
}
\arguments{
\item{w}{weight for mixture prior distribution}

\item{hr1}{first assumed true treatment effect on HR scale for prior distribution}

\item{hr2}{second assumed true treatment effect on HR scale for prior distribution}

\item{id1}{amount of information for hr1 in terms of number of events}

\item{id2}{amount of information for hr2 in terms of number of events}

\item{d2min}{minimal number of events for phase II}

\item{d2max}{maximal number of events for phase II}

\item{stepd2}{stepsize for the optimization over d2}

\item{hrgomin}{minimal threshold value for the go/no-go decision rule}

\item{hrgomax}{maximal threshold value for the go/no-go decision rule}

\item{stephrgo}{stepsize for the optimization over HRgo}

\item{alpha}{significance level}

\item{beta}{1-beta power for calculation of the number of events for phase III by Schoenfeld (1981) formula}

\item{xi2}{event rate for phase II}

\item{xi3}{event rate for phase III}

\item{c2}{variable per-patient cost for phase II}

\item{c3}{variable per-patient cost for phase III}

\item{c02}{fixed cost for phase II}

\item{c03}{fixed cost for phase III}

\item{K}{constraint on the costs of the program, default: Inf, e.g. no constraint}

\item{N}{constraint on the total expected sample size of the program, default: Inf, e.g. no constraint}

\item{S}{constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint}

\item{steps1}{lower boundary for effect size category "small" in HR scale, default: 1}

\item{stepm1}{lower boundary for effect size category "medium" in HR scale = upper boundary for effect size category "small" in HR scale, default: 0.95}

\item{stepl1}{lower boundary for effect size category "large" in HR scale = upper boundary for effect size category "medium" in HR scale, default: 0.85}

\item{b1}{expected gain for effect size category "small"}

\item{b2}{expected gain for effect size category "medium"}

\item{b3}{expected gain for effect size category "large"}

\item{gamma}{to model different populations in phase II and III choose gamma!=0, default: 0}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE hr1 is used as fixed effect}

\item{skipII}{choose if skipping phase II is an option, default: FASLE}

\item{num_cl}{number of clusters used for parallel computing, default: 1}
}
\value{
The output of the function \code{\link{optimal_tte}} is a data.frame containing the optimization results:
\describe{
  \item{u}{maximal expected utility}
  \item{HRgo}{optimal threshold value for the decision rule to go to phase III}
  \item{d2}{optimal total number of events for phase II}
  \item{d3}{total expected number of events for phase III; rounded to next natural number}
  \item{d}{total expected number of events in the program; d = d2 + d3}
  \item{n2}{total sample size for phase II; rounded to the next even natural number}
  \item{n3}{total sample size for phase III; rounded to the next even natural number}
  \item{n}{total sample size in the program; n = n2 + n3}
  \item{K}{maximal costs of the program}
  \item{pgo}{probability to go to phase III}
  \item{sProg}{probability of a successful program}
  \item{sProg1}{probability of a successful program with "small" treatment effect in Phase III}
  \item{sProg2}{probability of a successful program with "medium" treatment effect in Phase III}
  \item{sProg3}{probability of a successful program with "large" treatment effect in Phase III }
  \item{K2}{expected costs for phase II}
  \item{K3}{expected costs for phase III}
  }
and further input parameters.

Taking cat(comment()) of the data.frame object lists the used optimization sequences, start and finish date of the optimization procedure.
}
\description{
The drugdevelopR package enables planning of phase II/III drug development programs with optimal sample size allocation and go/no-go decision rules. For time-to-event endpoints the treatment effect is measured by the hazard ratio (HR) and more information on the framework can be found in Kirchner et al. (2016). The assumed true treatment effects can be assumed fixed or modelled by a prior distribution. The R Shiny application \href{https://web.imbi.uni-heidelberg.de/prior/}{prior} visualizes the prior distributions used in this package. Fast coputing is enabled by parallel programming.
}
\section{drugdevelopR functions}{

The drugdevelopR package provides the functions
\itemize{
  \item \code{\link{optimal_tte}},
  \item \code{\link{optimal_binary}} and
  \item \code{\link{optimal_normal}}
}
to plan optimal phase II/III drug development programs with
\itemize{
  \item time-to-event (treatment effect measured by hazard ratio (HR)),
  \item binary (treatment effect measured by risk ratio (RR)) or
  \item normally distributed (treatment effect measured by standardized difference in means (Delta))
}
endpoint, where the treatment effect is modelled by a \href{https://web.imbi.uni-heidelberg.de/prior/}{prior}. Optimal phase II/III drug development planning with fixed treatment effects can be done with the help of the R Shiny application \href{https://web.imbi.uni-heidelberg.de/basic/}{basic}. Extensions are 
\itemize{
  \item optimal planning of programs including methods for discounting of phase II results (function: \code{\link{optimal_bias}}, App: \href{https://web.imbi.uni-heidelberg.de/bias/}{bias}),
  \item optimal planning of programs with several phase III trials (function: \code{\link{optimal_multitrial}}, App: \href{https://web.imbi.uni-heidelberg.de/multitrial/}{multitrial}) and
  \item optimal planning of programs with multiple arms (function: \code{\link{optimal_multiarm}}, App: \href{https://web.imbi.uni-heidelberg.de/multiarm/}{multiarm}).
}
}

\examples{
res <- optimal_tte(w = 0.3,                              # define parameters for prior
  hr1 = 0.69, hr2 = 0.88, id1 = 210, id2 = 420,          # (https://web.imbi.uni-heidelberg.de/prior/)
  d2min = 20, d2max = 100, stepd2 = 5,                   # define optimization set for d2
  hrgomin = 0.7, hrgomax = 0.9, stephrgo = 0.05,         # define optimization set for HRgo
  alpha = 0.05, beta = 0.1, xi2 = 0.7, xi3 = 0.7,        # drug development planning parameters
  c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,               # define fixed and variable costs for phase II and III
  K = Inf, N = Inf, S = -Inf,                            # set maximal costs/ expected sample size for the program or minimal expected probability of a successful program
  steps1 = 1,                                            # define lower boundary for "small"
  stepm1 = 0.95,                                         # "medium"
  stepl1 = 0.85,                                         # and "large" treatment effect size categories as proposed by IQWiG (2016)
  b1 = 1000, b2 = 2000, b3 = 3000,                       # define expected benefit for a "small", "medium" and "large" treatment effect
  gamma = 0,                                             # assume different/same population structures in phase II and III
  fixed = FALSE,                                         # choose if true treatment effects are fixed or random
  skipII = FALSE,                                        # choose if skipping phase II would be an option
  num_cl = 1)                                            # set number of cores used for parallelized computing (check maximum number possible with detectCores())
res
cat(comment(res))                                        # displays the optimization sequence, start and finish date of the optimization procedure.
}
\references{
Kirchner, M., Kieser, M., Goette, H., & Schueler, A. (2016). Utility-based optimization of phase II/III programs. Statistics in Medicine, 35(2), 305-316.

IQWiG (2016). Allgemeine Methoden. Version 5.0, 10.07.2016, Technical Report. Available at \href{https://www.iqwig.de/de/methoden/methodenpapier.3020.html}{https://www.iqwig.de/de/methoden/methodenpapier.3020.html}, assessed last 15.05.19.

Schoenfeld, D. (1981). The asymptotic properties of nonparametric tests for comparing survival distributions. Biometrika, 68(1), 316-319.
}
\seealso{
\code{\link{optimal_binary}}
}
