% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_tte.R
\name{pgo_multipe_tte}
\alias{pgo_multipe_tte}
\title{probability to go to phase III: pgo}
\usage{
pgo_multipe_tte(HRgo, n2, ec, hr1, hr2, id1, id2, fixed, rho)
}
\arguments{
\item{HRgo}{threshold value for the go/no-go decision rule;}

\item{n2}{total sample size for phase II; must be even number}

\item{ec}{control arm event rate for phase II and III}

\item{hr1}{assumed true treatment effect on HR scale for treatment 1}

\item{hr2}{assumed true treatment effect on HR scale for treatment 2}

\item{id1}{amount of information for hr1 in terms of number of events}

\item{id2}{amount of information for hr2 in terms of number of events}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE hr1 is used as fixed effect}

\item{rho}{correlation between the two endpoints}
}
\value{
the output of the the function pgo_multiple_tte is the probability to go to phase III
}
\description{
probability to go to phase III: pgo
}
\examples{
res <- pgo_multiple_tte(HRgo = 0.8, n2 = 50, ec = 0.6,
                               hr1 = 0.75, hr2 = 0.80, id1 = 300, id2 = 600, 
                               fixed = FALSE, rho = 0.3)
}
