% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_normal.R
\name{pgo_multiple_normal}
\alias{pgo_multiple_normal}
\title{probability to go to phase III: pgo}
\usage{
pgo_multiple_normal(
  kappa,
  n2,
  Delta1,
  Delta2,
  in1,
  in2,
  sigma1,
  sigma2,
  fixed,
  rho
)
}
\arguments{
\item{kappa}{threshold value for the go/no-go decision rule; vector for both endpoints}

\item{n2}{total sample size for phase II; must be even number}

\item{Delta1}{assumed true treatment effect for standardized difference in means}

\item{Delta2}{assumed true treatment effect for standardized difference in means}

\item{in1}{amount of information for Delta1 in terms of sample size}

\item{in2}{amount of information for Delta2 in terms of sample size}

\item{sigma1}{standard deviation of first endpoint}

\item{sigma2}{standard deviation of second endpoint}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE Delta1 is used as fixed effect}

\item{rho}{correlation between the two endpoints}
}
\value{
the output of the the function pgo_multiple_normal is the probability to go to phase III
}
\description{
probability to go to phase III: pgo
}
\examples{
res <- pgo_multiple_normal(kappa = c(0.1,0.1), n2 = 50,
                               Delta1 = 0.375, Delta2 = 0.625, in1 = 300, in2 = 600, 
                               sigma1 = 8, sigma2 = 4, fixed = FALSE, rho = 0.3)
}
