% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_tte.R
\name{pw}
\alias{pw}
\title{probabilty that P(x>y)=P(x-y>0)
Z=X-Y is normally distributed with expectation mux-muy and variance sigmax+sigmay-2 rho sdx sdy}
\usage{
pw(n2, ec, hr1, hr2, id1, id2, fixed, rho)
}
\arguments{
\item{n2}{total sample size for phase II; must be even number}

\item{ec}{control arm event rate for phase II and III}

\item{hr1}{assumed true treatment effect on HR scale for treatment 1}

\item{hr2}{assumed true treatment effect on HR scale for treatment 2}

\item{id1}{amount of information for hr1 in terms of number of events}

\item{id2}{amount of information for hr2 in terms of number of events}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE hr1 is used as fixed effect}

\item{rho}{correlation between the two endpoints}
}
\value{
the output of the the function pw is the the probabilty that endpoint 1 has a better result than endpoint two
}
\description{
probabilty that P(x>y)=P(x-y>0)
Z=X-Y is normally distributed with expectation mux-muy and variance sigmax+sigmay-2 rho sdx sdy
}
\examples{
res <- pw(n2 = 50, ec = 0.6,
                    hr1 = 0.75, hr2 = 0.80, id1 = 300, id2 = 600, 
                    fixed = FALSE, rho = 0.3)
}
