% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_tte.R
\name{utility_multiple_tte}
\alias{utility_multiple_tte}
\title{expected probability of a successful program}
\usage{
utility_multiple_tte(
  n2,
  HRgo,
  alpha,
  beta,
  hr1,
  hr2,
  id1,
  id2,
  ec,
  c2,
  c02,
  c3,
  c03,
  K,
  N,
  S,
  steps1,
  stepm1,
  stepl1,
  b1,
  b2,
  b3,
  fixed,
  rho
)
}
\arguments{
\item{n2}{total sample size for phase II; must be even number}

\item{HRgo}{threshold value for the go/no-go decision rule;}

\item{alpha}{significance level}

\item{beta}{1-beta power for calculation of sample size for phase III}

\item{hr1}{assumed true treatment effect on HR scale for treatment 1}

\item{hr2}{assumed true treatment effect on HR scale for treatment 2}

\item{id1}{amount of information for hr1 in terms of sample size}

\item{id2}{amount of information for hr2 in terms of sample size}

\item{ec}{control arm event rate for phase II and III}

\item{c2}{variable per-patient cost for phase II}

\item{c02}{fixed cost for phase II}

\item{c3}{variable per-patient cost for phase III}

\item{c03}{fixed cost for phase III}

\item{K}{constraint on the costs of the program, default: Inf, e.g. no constraint}

\item{N}{constraint on the total expected sample size of the program, default: Inf, e.g. no constraint}

\item{S}{constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint}

\item{steps1}{lower boundary for effect size category "small" in HR scale, default: 1}

\item{stepm1}{lower boundary for effect size category "medium" in HR scale = upper boundary for effect size category "small" in HR scale, default: 0.95}

\item{stepl1}{lower boundary for effect size category "large" in HR scale = upper boundary for effect size category "medium" in HR scale, default: 0.85}

\item{b1}{expected gain for effect size category "small"}

\item{b2}{expected gain for effect size category "medium"}

\item{b3}{expected gain for effect size category "large"}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE hr1 is used as fixed effect}

\item{rho}{correlation between the two endpoints}
}
\value{
the output of the the function utility_multiple_tte is the expected probability of a successfull program, when going to phase III
}
\description{
expected probability of a successful program
}
\examples{
res <- utility_multiple_tte(n2 = 50, HRgo = 0.8, alpha = 0.025, beta = 0.1,
                               hr1 = 0.75, hr2 = 0.80,
                               id1 = 300, id2 = 600, ec = 0.6,
                               c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,               
                               K = Inf, N = Inf, S = -Inf, 
                               steps1 = 1, stepm1 = 0.95, stepl1 = 0.85,
                               b1 = 1000, b2 = 2000, b3 = 3000,
                               fixed = FALSE, rho = 0.3)
}
