% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_normal.R,
%   R/functions_multiple_tte.R
\name{dbivanorm}
\alias{dbivanorm}
\title{density of bivariate normal distribution}
\usage{
dbivanorm(x, y, mu1, mu2, sigma1, sigma2, rho)

dbivanorm(x, y, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{x}{integral variable}

\item{y}{integral variable}

\item{mu1}{mean of second endpoint}

\item{mu2}{mean of first endpoint}

\item{sigma1}{standard deviation of first endpoint}

\item{sigma2}{standard deviation of second endpoint}

\item{rho}{correlation between the two endpoints}
}
\value{
function dbivanrom will return the density of bivariate normal distribution

function dbivanrom will return the density of bivariate normal distribution
}
\description{
density of bivariate normal distribution

density of bivariate normal distribution
}
\examples{
res <- fmin(x = 0.5, y = 0.5, mu1 = 0.375, mu2 = 0.25, sigma1 = 8, sigma2 = 12, rho = 0.4 )
res <- fmin(x = 0.5, y = 0.5, mu1 = 0.375, mu2 = 0.25, sigma1 = 8, sigma2 = 12, rho = 0.4 )
}
