% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multitrial.R
\name{Epgo23}
\alias{Epgo23}
\title{Expected probability to do third phase III trial}
\usage{
Epgo23(HRgo, d2, alpha, beta, w, hr1, hr2, id1, id2)
}
\arguments{
\item{HRgo}{threshold value for the go/no-go decision rule}

\item{d2}{total number of events in phase II}

\item{w}{weight for mixture prior distribution}

\item{hr1}{first assumed true treatment effect on HR scale for prior distribution}

\item{hr2}{second assumed true treatment effect on HR scale for prior distribution}

\item{id1}{amount of information for \code{hr1} in terms of number of events}

\item{id2}{amount of information for \code{hr2} in terms of number of events}

\item{p0}{assumed true rate of control group}
}
\value{
The output of the the function \code{Epgo23()} is the probability to to a third phase III trial.
}
\description{
In the setting of Case 2: Strategy 2/2( + 1); at least two trials significant (and the
treatment effect of the other one at least showing in the same direction) this function calculates the probability that a third phase III trial is necessary.
}
\examples{
res <- Epgo23(HRgo = 0.8, d2 = 50,  w = 0.3,
                               hr1 =  0.69, hr2 = 0.81, id1 = 280, id2 = 420)
}
\section{Last Edited By}{
Johannes Cepicka
}

\section{Last Edit Date}{
2022-05-09
}

