% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multitrial_normal.R
\name{EPsProg23_normal}
\alias{EPsProg23_normal}
\title{Expected probability of a successful program, when 2 or 3 phase phase III trials are performed (outcomes normally distributed)}
\usage{
EPsProg23_normal(
  kappa,
  n2,
  alpha,
  beta,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  case,
  size,
  ymin
)
}
\arguments{
\item{kappa}{threshold value for the go/no-go decision rule}

\item{n2}{total sample size for phase II; must be even number}

\item{alpha}{significance level}

\item{beta}{1-beta power for calculation of sample size for phase III}

\item{w}{weight for mixture prior distribution}

\item{Delta1}{assumed true treatment effect for standardized difference in means}

\item{Delta2}{assumed true treatment effect for standardized difference in means}

\item{in1}{amount of information for Delta1 in terms of sample size}

\item{in2}{amount of information for Delta2 in terms of sample size}

\item{a}{lower boundary for the truncation}

\item{b}{upper boundary for the truncation}

\item{case}{choose case: "at least 1, 2 or 3 significant trials needed for approval"}

\item{size}{effect size category "small", "medium", "large" or "all"}

\item{fixed}{choose if true treatment effects are fixed or random}
}
\value{
the output of the the function EPsProg23_normal is the expected probability of a successful program
}
\description{
Option 2.1: first two phase III trials are successful: no third phase III trial
Option 2.2: one of the two first phase III trials successful, the treatment
effect of the other one points in the same direction: 
conduct third phase III trial with N3 = N3(ymin)
}
\examples{
res <- EPsProg23_normal(kappa = 0.1, n2 = 50 alpha = 0.025, beta = 0.1, w = 0.3,
                                 Delta1 = 0.375, Delta2 = 0.625, in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, 
                                 case = 2, size = "small", ymin = 0.5)
}
