% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiple_tte.R
\name{os_tte}
\alias{os_tte}
\title{Probability that endpoint OS significant}
\usage{
os_tte(HRgo, n2, alpha, beta, hr1, hr2, id1, id2, fixed, rho)
}
\arguments{
\item{HRgo}{threshold value for the go/no-go decision rule;}

\item{n2}{total sample size for phase II; must be even number}

\item{alpha}{one- sided significance level}

\item{beta}{1-beta power for calculation of the number of events for phase III by Schoenfeld (1981) formula}

\item{hr1}{assumed true treatment effect on HR scale for treatment 1}

\item{hr2}{assumed true treatment effect on HR scale for treatment 2}

\item{id1}{amount of information for \code{hr1} in terms of number of events}

\item{id2}{amount of information for \code{hr2} in terms of number of events}

\item{fixed}{choose if true treatment effects are fixed or random, if TRUE \code{hr1} is used as fixed effect}

\item{rho}{correlation between the two endpoints}
}
\value{
The output of the the function \code{os_tte()} is the probability that endpoint OS significant.
}
\description{
This function calculate the probability that the endpoint OS is statistically significant.
In the context of cancer research OS stands for overall survival, a positive treatment effect in this endpoints is thus sufficient for a successful program.
}
\examples{
res <- os_tte(HRgo = 0.8, n2 = 50, alpha = 0.05, beta = 0.1,
                               hr1 = 0.75, hr2 = 0.80, 
                               id1 = 300, id2 = 600, 
                               fixed = TRUE, rho = 0.3)
}
