% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.drugdevelopResult.R
\name{print_drugdevelopResult_helper}
\alias{print_drugdevelopResult_helper}
\title{Helper function for printing a drugdevelopResult Object}
\usage{
print_drugdevelopResult_helper(x, ...)
}
\arguments{
\item{x}{Data frame}

\item{...}{Further arguments.}
}
\value{
No return value, called for printing to the console using \code{cat()}
}
\description{
Helper function for printing a drugdevelopResult Object
}
\keyword{internal}
